requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.7.0
click==8.1.3
Django==4.1.1
graphql-core==3.2.1
h11==0.13.0
idna==3.4
psycopg2==2.9.3
pycodestyle==2.9.1
Pygments==2.13.0
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.20.4
strawberry-django-plus==1.26.1
strawberry-graphql==0.131.1
strawberry-graphql-django==0.5.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.3
